+++ src/app.module.ts
@@ -16,7 +16,8 @@ import { BatchModule } from './batch/batch.module';
 import dbConfig from './config/db.config';
 import serverConfig from './config/server.config';
 import { NodeEnv } from './shared/enum/node-env.enum';
-import { AllExceptionFilter, HttpExceptionFilter } from './shared/filter/custom-exception.filter';
+import { AllExceptionFilter } from './shared/filter/all-exception.filter';
+import { HttpExceptionFilter } from './shared/filter/http-exception.filter';
 import { TransformInterceptor } from './shared/interceptor/transform.interceptor';
 import { LoggingInterceptor } from './shared/interceptor/logging.interceptor';
 import { DbConfig, ServerConfig } from './shared/interface/config.interface';
